# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "description": "Chỉ dành cho Admin",
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo thành công"
          },
          "400": {
            "description": "Dữ liệu đầu vào không hợp lệ"
          },
          "403": {
            "description": "Không có quyền truy cập"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Tạo tài khoản mới",
        "tags": [
          "Người dùng (Users)"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Người dùng (Users)"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Người dùng (Users)"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Người dùng (Users)"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Người dùng (Users)"
        ]
      }
    },
    "/whitelist-ip": {
      "post": {
        "operationId": "WhitelistIpController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWhitelistIpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "WhitelistIp"
        ]
      },
      "get": {
        "operationId": "WhitelistIpController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "WhitelistIp"
        ]
      }
    },
    "/whitelist-ip/{id}": {
      "get": {
        "operationId": "WhitelistIpController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "WhitelistIp"
        ]
      },
      "patch": {
        "operationId": "WhitelistIpController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWhitelistIpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "WhitelistIp"
        ]
      },
      "delete": {
        "operationId": "WhitelistIpController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "WhitelistIp"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS Pro API",
    "description": "Tài liệu API chi tiết cho dự án",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local environment"
    },
    {
      "url": "https://yourdomain.com",
      "description": "Staging"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Nhập mã token vào đây",
        "in": "header"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "@description Họ và tên đầy đủ",
            "example": "Nguyễn Văn A",
            "maxLength": 50
          },
          "username": {
            "type": "string",
            "description": "Tên đăng nhập (chỉ chứa chữ, số và _)",
            "example": "admin_pro",
            "minLength": 5,
            "pattern": "^[a-zA-Z0-9_]+$"
          },
          "password": {
            "type": "string",
            "description": "Mật khẩu (ít nhất 1 hoa, 1 thường, 1 số)",
            "example": "Password@123",
            "minLength": 8
          }
        },
        "required": [
          "name",
          "username",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
      "CreateWhitelistIpDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateWhitelistIpDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}